Refactor CSRF feature to implement ServerRestHandler
