+++ extensions/csrf-reactive/runtime/src/main/java/io/quarkus/csrf/reactive/runtime/CsrfRecorder.java
@@ -1,0 +1,19 @@
+package io.quarkus.csrf.reactive.runtime;
+
+import java.util.function.Supplier;
+
+import io.quarkus.runtime.annotations.Recorder;
+
+@Recorder
+public class CsrfRecorder {
+
+    public Supplier<CsrfReactiveConfig> configure(CsrfReactiveConfig csrfReactiveConfig) {
+        return new Supplier<CsrfReactiveConfig>() {
+            @Override
+            public CsrfReactiveConfig get() {
+                return csrfReactiveConfig;
+            }
+        };
+    }
+
+}
