+++ integration-tests/csrf-reactive/src/main/resources/templates/csrfTokenMultipart.html
@@ -11,6 +11,7 @@
     	<input type="hidden" name="{inject:csrf.parameterName}" value="{inject:csrf.token}" />
     	
     	<p>Your Name: <input type="text" name="name" /></p>
+    	<p>Your Name: <input type="file" name="file" /></p>
     	<p><input type="submit" name="submit"/></p>
     </form>
 </body>
